pyproject.toml
[tool.poetry]
name = "poetry-mongodb"
version = "0.1.0"
description = ""
authors = ["Yuchen <yucpeng17@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pymongo = "^4.8.0"
pyyaml = "^6.0.1"
pandas = "^2.2.2"
requests = "^2.32.3"
bs4 = "^0.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
